Allow exporting function as shorthand
Depends on #5 being pulled first.

Let's you just do

    export default function(container, app) {
      // blah blah
    };

Just one less thing to have to remember. 
I considered adding this to core Ember, but
the problem is that you need to a supply a name
for the DAG, which we can do in EAK/CLI. I 
suppose we could make it so that an initializer's
default name in ember core is a GUID, but I 
think it's fine for us to add this to ember-cli
first, see how it feels.